D2LScrape/requirements.txt
PyPDF2>=3.0.0
Pillow>=9.0.0
pdf2image>=1.16.0
pytesseract>=0.3.8
sentence-transformers
numpy
torch
tqdm 